Fix SQLite MarketTransaction decimal columns (REAL to TEXT)
- Bump schema version from 2 to 3
- Add MigrateMarketTransactionsDecimalToText migration (v2->v3) that rebuilds
  the MarketTransactions table with TEXT columns for PricePerUnit, TotalPrice,
  and MaxRepairPercent
- Update initial schema: PricePerUnit, TotalPrice, MaxRepairPercent from REAL to TEXT
- Remove (double) casts in UpsertMarketTransactionAsync, use .ToString(G)
- Update ReadMarketTransaction to use decimal.Parse with CultureInfo.InvariantCulture

Satisfies: storage-migration-integration Req 7, Criterion 2 (task 1.2)

Prompt: Execute task 1.2 from the storage-migration-integration spec.

+++ OE2EmpireTracker.Common/Storage/PostgresBackend.cs
@@ -2789,8 +2789,8 @@ CREATE TABLE IF NOT EXISTS PropertyTypeDefinitions (
                     cmd.Parameters.AddWithValue("@itemRefId", entity.ItemReferenceID ?? string.Empty);
                     cmd.Parameters.AddWithValue("@itemName", entity.ItemName ?? string.Empty);
                     cmd.Parameters.AddWithValue("@qty", entity.Quantity);
-                    cmd.Parameters.AddWithValue("@price", (double)entity.PricePerUnit);
-                    cmd.Parameters.AddWithValue("@total", (double)entity.TotalPrice);
+                    cmd.Parameters.AddWithValue("@price", entity.PricePerUnit);
+                    cmd.Parameters.AddWithValue("@total", entity.TotalPrice);
                     cmd.Parameters.AddWithValue("@counterparty", entity.Counterparty ?? string.Empty);
                     cmd.Parameters.AddWithValue("@counterpartyFaction", entity.CounterpartyFaction ?? string.Empty);
                     cmd.Parameters.AddWithValue("@stationUUID", entity.StationUUID ?? string.Empty);
@@ -2799,7 +2799,7 @@ CREATE TABLE IF NOT EXISTS PropertyTypeDefinitions (
                     cmd.Parameters.AddWithValue("@listingUUID", entity.ListingUUID ?? string.Empty);
                     cmd.Parameters.AddWithValue("@curHp", entity.CurrentHP);
                     cmd.Parameters.AddWithValue("@maxHp", entity.MaxHP);
-                    cmd.Parameters.AddWithValue("@maxRepair", (double)entity.MaxRepairPercent);
+                    cmd.Parameters.AddWithValue("@maxRepair", entity.MaxRepairPercent);
                     cmd.ExecuteNonQuery();
                 }
             });
